PER-135: move category/tag browse out of the nav into a subtle row + filter
The nav felt bloated with Categories/Tags. Moved them to a thin, understated
BrowseBar between the title bar and the content (subtle uppercase mono text),
keeping the nav lean (kp~ / Blog Wiki About / theme toggle / RSS).

- components/BrowseBar.tsx: Categories + Tags dropdowns (from GlobalContext),
  rendered by PageShell below SiteTitle. The Tags dropdown gets a filter input
  (shown when >8 items) so the long tag list is searchable.
- SiteHeader: dropdowns removed; back to a lean nav.

Verified: full gate PASSES (23 Playwright incl. a11y); prod-build screenshot
shows the lean nav + browse row + working tag filter.
